requirements.txt
python-telegram-bot==20.3
Flask==2.2.5
requests==2.31.0